Using `__dict__` to serialize objects improve the overall noop query latency

Rough benchmark

before
  25% in 0.0054 secs
  50% in 0.0060 secs
  75% in 0.0066 secs
  90% in 0.0068 secs

after
  25% in 0.0044 secs
  50% in 0.0046 secs
  75% in 0.0048 secs
  90% in 0.0059 secs

<!-- Thank you for your contribution! Please review https://github.com/ray-project/ray/blob/master/CONTRIBUTING.rst before opening a pull request. -->

## Why are these changes needed?

<!-- Please give a short summary of the change and the problem this solves. -->

## Related issue number

<!-- For example: "Closes #1234" -->

## Checks

- [x] I've run `scripts/format.sh` to lint the changes in this PR.
- [ ] I've included any doc changes needed for https://ray.readthedocs.io/en/latest/.
- [ ] I've made sure the tests are passing. Note that there might be a few flaky tests, see the recent failure rates at https://ray-travis-tracker.herokuapp.com/.
- Testing Strategy
   - [ ] Unit tests
   - [ ] Release tests
   - [ ] This PR is not tested (please justify below)